refactor: deduplicate replay utilities, simplify onOpenReplay, remove stale key handler
- Extract shared replay playback utilities (formatReplayTime, clamp, targetKey,
  toFrameTarget, PLAYBACK_SPEEDS, etc.) to src/components/replay-utils.ts
- Update ReplayTerminal.tui.tsx and ReplayTerminal.web.tsx to import from shared module
- Simplify onOpenReplay from (params: { replayId, workspaceId }) to (replayId: string)
  since workspaceId was never used by any handler
- Remove orphaned 'h' keydown listener that fired in web replay view, confusingly
  toggling the dismissed-replay filter while the filter wasn't visible
