+++ src/components/ReplayTerminal.web.tsx
@@ -1,11 +1,16 @@
 /** @jsxImportSource react */
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import type { ReplayFrameTarget, ReplayInfo, ReplayTimeline, ReplayTimelineStep } from '../lib/tmux-lite/replay/index.js';
+import type { ReplayFrameTarget, ReplayInfo, ReplayTimeline } from '../lib/tmux-lite/replay/index.js';
 import { SessionTerminal } from './SessionTerminal.web';
-
-const PLAYBACK_SPEEDS = [0.25, 0.5, 1, 1.5, 2, 4] as const;
-const DEFAULT_PLAYBACK_SPEED_INDEX = 2;
-const FAST_SCRUB_STEP_COUNT = 5;
+import {
+  PLAYBACK_SPEEDS,
+  DEFAULT_PLAYBACK_SPEED_INDEX,
+  FAST_SCRUB_STEP_COUNT,
+  formatReplayTime,
+  clamp,
+  targetKey,
+  toFrameTarget,
+} from './replay-utils.js';
 
 function encodeAnsi(text: string): Uint8Array {
   return new TextEncoder().encode(text);
@@ -13,46 +18,6 @@ function encodeAnsi(text: string): Uint8Array {
 
 const LOADING_REPLAY_ANSI = encodeAnsi('\x1b[2J\x1b[H\x1b[2;37mLoading replay...\x1b[0m');
 const EMPTY_REPLAY_ANSI = encodeAnsi('\x1b[2J\x1b[H\x1b[2;37m(empty replay)\x1b[0m');
-
-function formatReplayTime(timeMs: number): string {
-  const totalSeconds = Math.max(0, timeMs) / 1000;
-  if (totalSeconds < 59.95) {
-    return `${totalSeconds.toFixed(1)}s`;
-  }
-
-  const wholeSeconds = Math.round(totalSeconds);
-  const seconds = wholeSeconds % 60;
-  const minutes = Math.floor(wholeSeconds / 60) % 60;
-  const hours = Math.floor(wholeSeconds / 3600);
-
-  if (hours > 0) {
-    return `${hours}:${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`;
-  }
-
-  return `${minutes}:${String(seconds).padStart(2, '0')}`;
-}
-
-function clamp(value: number, min: number, max: number): number {
-  return Math.min(max, Math.max(min, value));
-}
-
-function targetKey(target: ReplayFrameTarget | ReplayTimelineStep | null): string | null {
-  if (!target) {
-    return null;
-  }
-
-  const timeMs = 'timeMs' in target ? target.timeMs : target.atMs;
-  const seq = 'seq' in target ? target.seq : target.atSeq;
-  return `${timeMs ?? -1}:${seq ?? -1}`;
-}
-
-function toFrameTarget(step: ReplayTimelineStep | null, fallback: ReplayFrameTarget): ReplayFrameTarget {
-  if (!step) {
-    return fallback;
-  }
-
-  return { atMs: step.timeMs, atSeq: step.seq };
-}
 
 export interface ReplayTerminalWebProps {
   replay: ReplayInfo;
